Changes to follow db rules and started working on rust db access scripts

+++ yt_downlaoder/yt_download_manager/src/lib.rs
@@ -11,7 +11,7 @@ pub mod models {
         pub url: String,
         pub thumbnail_url: String,
         pub video_id: String,
-        pub size: u64,
+        pub size: i32,
     }
     impl From<web::Json<FetchMeta>> for FetchMeta {
         fn from(fetched: web::Json<FetchMeta>) -> Self {
